All grammar paths tested. They work.
Addition, Subtraction, Multiplication, and Division are all RIGHT ASSOCIATIVE. Why? No real clue, it was part of the project as far as I can tell. Its how the parse tree is built. It feels wrong, but it would have to be intentionally put in that way back in P2.

VM spec added.